# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: KACP API
  version: 0.1.0-draft
  description: |
    KCC Agent Cloud Platform 외부 API (0단계 초안).
    규칙과 내부 API는 docs/design/04-api.md 참고.
    상태 변경 요청은 X-KACP-CSRF 헤더가 필요하다.
servers:
  - url: https://app.kacp.cloud/api/v1
security:
  - session: []
...
paths:
...
  /admin/settings:
    get:
      tags: [admin]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlatformSettings' }
    put:
      tags: [admin]
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlatformSettings' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlatformSettings' }
  /admin/settings/api-keys/{provider}:
    parameters:
      - { name: provider, in: path, required: true, schema: { type: string } }
    put:
      tags: [admin]
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key: { type: string }
      responses:
        '204': { description: 저장됨 }
  /admin/audit-events:
    get:
      tags: [admin]
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: actor, in: query, schema: { type: string, format: uuid } }
        - { name: action, in: query, schema: { type: string } }
        - { name: targetType, in: query, schema: { type: string } }
        - { name: targetId, in: query, schema: { type: string }, description: "A-03 '이 사람에게 일어난 일'(targetType=user)" }
        - { name: team, in: query, schema: { type: string } }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/AuditEvent' }
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: kacp_session
  parameters:
    Csrf:
      name: X-KACP-CSRF
      in: header
      required: true
      schema: { type: string }
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
...
  schemas:
...
    Page:
      type: object
      properties:
        nextCursor: { type: [string, 'null'] }
...
    UserRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string }
        departmentName: { type: [string, 'null'], description: 동명이인 구분용 표시 }
...
    ResourceLimits:
      type: object
      properties:
        cpu: { type: number, description: vCPU }
        memoryMb: { type: integer }
        diskGb: { type: integer }
...
    PlatformSettings:
      type: object
      properties:
        models:
          type: object
          properties:
            allowed:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  label: { type: string }
                  provider: { type: string }
                  default: { type: boolean }
            apiKeysConfigured:
              type: object
              additionalProperties: { type: boolean }
              readOnly: true
        limits:
          type: object
          properties:
            teamDefault: { $ref: '#/components/schemas/ResourceLimits' }
            appDefault: { $ref: '#/components/schemas/ResourceLimits' }
            mcpDefault: { $ref: '#/components/schemas/ResourceLimits' }
        ops:
          type: object
          properties:
            idleStopMinutes: { type: integer, description: 팀 컨테이너 }
            appIdleStopMinutes:
              type: object
              properties:
                work: { type: integer }
                public: { type: integer }
            maxRunningWorkAppsPerTeam: { type: integer }
            capacityWarn:
              type: object
              description: A-01 노랑(warn)·빨강(danger) 두 단계 (03-data-model ops.capacity_warn)
              properties:
                warn: { $ref: '#/components/schemas/CapacityThresholds' }
                danger: { $ref: '#/components/schemas/CapacityThresholds' }
            trashRetentionDays: { type: integer }
        names:
          type: object
          properties:
            reserved: { type: array, items: { type: string }, readOnly: true }
            reservedExtra: { type: array, items: { type: string } }
    CapacityThresholds:
      type: object
      properties:
        cpu: { type: integer }
        memory: { type: integer }
        disk: { type: integer }
    AuditEvent:
      type: object
      properties:
        id: { type: integer }
        at: { type: string, format: date-time }
        actor: { oneOf: [{ $ref: '#/components/schemas/UserRef' }, { type: 'null' }] }
        action: { type: string }
        targetType: { type: string }
        targetId: { type: string }
        targetLabel: { type: string }
        team: { type: [string, 'null'] }
        detail: { type: object }
        ip: { type: [string, 'null'] }